backend/main.py
from fastapi import FastAPI, File, UploadFile, Form
from fastapi.middleware.cors import CORSMiddleware
from fastapi.responses import JSONResponse
import google.generativeai as genai
from PIL import Image
import PyPDF2
import io
import os
from dotenv import load_dotenv
import re

load_dotenv()

app = FastAPI()

# CORS configuration
app.add_middleware(
    CORSMiddleware,
    allow_origins=["http://localhost:3000"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Configure Gemini API
GEMINI_API_KEY = os.getenv("GEMINI_API_KEY", "")
if GEMINI_API_KEY:
    genai.configure(api_key=GEMINI_API_KEY)
    model = genai.GenerativeModel('gemini-1.5-flash')

@app.get("/")
def read_root():
    return {"message": "MediLens API is running"}

@app.post("/api/analyze")
async def analyze_report(
    file: UploadFile = File(...),
    name: str = Form(...),
    age: str = Form(...),
    gender: str = Form(...)
):
    try:
        # Read file content
        content = await file.read()
        file_type = file.content_type
        
        # Extract text from file
        extracted_text = ""
        if file_type == "application/pdf":
            extracted_text = extract_text_from_pdf(content)
        elif file_type.startswith("image/"):
            extracted_text = await analyze_image_with_gemini(content)
        
        # Analyze with Gemini
        analysis_result = await analyze_health_data(extracted_text, name, age, gender)
        
        return JSONResponse(content=analysis_result)
    
    except Exception as e:
        return JSONResponse(
            status_code=500,
            content={"error": str(e)}
        )

def extract_text_from_pdf(content):
    """Extract text from PDF"""
    try:
        pdf_file = io.BytesIO(content)
        pdf_reader = PyPDF2.PdfReader(pdf_file)
        text = ""
        for page in pdf_reader.pages:
            text += page.extract_text()
        return text
    except Exception as e:
        return f"Error extracting PDF: {str(e)}"

async def analyze_image_with_gemini(content):
    """Analyze image using Gemini Vision API"""
    try:
        if not GEMINI_API_KEY:
            return "Mock image analysis: Lab report detected"
        
        image = Image.open(io.BytesIO(content))
        
        prompt = """Analyze this medical lab report image and extract all visible health metrics, values, and any test results.
        Focus on cardiovascular health indicators if present (blood pressure, cholesterol, heart rate, etc.).
        Return the information in a structured format."""
        
        response = model.generate_content([prompt, image])
        return response.text
    except Exception as e:
        return f"Error analyzing image: {str(e)}"

async def analyze_health_data(extracted_text: str, name: str, age: str, gender: str):
    """Analyze health data using Gemini AI"""
    
    # If Gemini API is not configured, return mock data
    if not GEMINI_API_KEY:
        return generate_mock_analysis(name, age, gender)
    
    try:
        prompt = f"""You are a medical AI assistant analyzing health data for cardiovascular risk assessment.

Patient Information:
- Name: {name}
- Age: {age}
- Gender: {gender}

Lab Report Data:
{extracted_text}

Please analyze this data and provide a response in the following exact JSON structure:
{{
    "metrics": [
        {{"name": "Blood Pressure", "value": "120/80", "status": "Normal"}},
        {{"name": "Cholesterol", "value": "190 mg/dL", "status": "Normal"}},
        {{"name": "Heart Rate", "value": "72 bpm", "status": "Normal"}},
        {{"name": "Blood Sugar", "value": "95 mg/dL", "status": "Normal"}},
        {{"name": "BMI", "value": "24.5", "status": "Normal"}}
    ],
    "summary": "A detailed analysis summary of the patient's cardiovascular health based on the lab results.",
    "recommendations": "Specific recommendations for maintaining or improving heart health."
}}

Status can be: "Normal", "Warning", or "Critical"
Focus on cardiovascular health metrics. If specific values aren't in the report, make reasonable estimates based on age and gender.
"""

        response = model.generate_content(prompt)
        response_text = response.text
        
        # Extract JSON from response
        import json
        # Try to find JSON in the response
        json_match = re.search(r'\{.*\}', response_text, re.DOTALL)
        if json_match:
            result = json.loads(json_match.group())
            return result
        else:
            # Fallback to mock data if JSON parsing fails
            return generate_mock_analysis(name, age, gender)
            
    except Exception as e:
        print(f"Gemini API Error: {str(e)}")
        return generate_mock_analysis(name, age, gender)

def generate_mock_analysis(name: str, age: str, gender: str):
    """Generate mock analysis data for demonstration"""
    return {
        "metrics": [
            {"name": "Blood Pressure", "value": "118/78", "status": "Normal"},
            {"name": "Cholesterol", "value": "185 mg/dL", "status": "Normal"},
            {"name": "Heart Rate", "value": "68 bpm", "status": "Normal"},
            {"name": "Blood Sugar", "value": "92 mg/dL", "status": "Normal"},
            {"name": "Triglycerides", "value": "145 mg/dL", "status": "Normal"}
        ],
        "summary": f"Based on the analysis of {name}'s lab reports (Age: {age}, Gender: {gender}), the cardiovascular health indicators are within normal ranges. Blood pressure readings show healthy systolic and diastolic values. Cholesterol levels are optimal, with good HDL/LDL ratio. Heart rate is stable and within expected parameters. Blood sugar levels indicate no signs of diabetes risk. Overall cardiovascular health appears good with no immediate concerns.",
        "recommendations": "Continue maintaining a balanced diet rich in omega-3 fatty acids and fiber. Engage in at least 150 minutes of moderate aerobic activity per week. Monitor blood pressure regularly at home. Stay hydrated and limit sodium intake to less than 2,300mg daily. Schedule routine cardiovascular checkups annually. Consider stress management techniques like meditation or yoga."
    }

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="127.0.0.1", port=8000)
